requirements.txt
Django==5.0
ipython==8.18.1
djangorestframework~=3.14.0
python-dotenv~=1.0.0
django-filter~=23.5
requests~=2.31.0
celery~=5.3.6